Add parameter tests to the DataQualityOperator
test: Add parameter tests to the DataQualityOperator according to the project reviewer

The operator was hard-coded but it should be parametrized. The operator was expected to get the test, the expected result as inputs and runs the tests dynamically based on the inputs. To overcome this, I passed a list of dicts as parameter to the operator. each dict has 2 values; the test and the expected result.