Updated parameter and for loop logic
test:  Update parameter and loop to convert the operator into testing engine and not a hard coded one.

The operator was hard-coded but it should be parametrized. The operator was expected to get the test, the expected result as inputs and runs the tests dynamically based on the inputs. To overcome this, I passed a list of dicts as parameter to the operator. each dict has 2 values; the test and the expected result. Then,  I loop over the list of dicts, executeed the test and compared the result to the expected value.